+++ src/main/kotlin/icu/windea/pls/PlsConstants.kt
@@ -100,7 +100,7 @@ val definitionIcon = IconLoader.getIcon("/icons/paradoxDefinition.svg",locationC
 val definitionLocalisationIcon = IconLoader.getIcon("/icons/paradoxDefinitionLocalisation.svg",locationClass)
 val localisationIcon = IconLoader.getIcon("/icons/paradoxLocalisation.svg",locationClass)
 val enumIcon = IconLoader.getIcon("/icons/paradoxEnum.svg",locationClass)
-val aliasIcon = IconLoader.getIcon("/icons/paradoxAlias.svg",locationClass)
+val aliasIcon = IconLoader.getIcon("/icons/paradoxAlias.svg",locationClass) //目前用不到，因为名字是表达式
 
 val definitionGutterIcon = definitionIcon.resize(12)
 val definitionLocalisationGutterIcon = definitionLocalisationIcon.resize(12)